server/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
googlemaps
pydantic
websockets
python-multipart
